+++ stock-system/src/main/java/saffchen/command/GenerateReportCommand.java
@@ -1,14 +1,31 @@
 package saffchen.command;
 
-public class GenerateReportCommand implements Command {
-    private ReceiverDB receiverDB;
+import saffchen.reports.PDFReportFromFileBySatellite;
 
-    public GenerateReportCommand(ReceiverDB receiverDB) {
-        this.receiverDB = receiverDB;
+import java.util.Scanner;
+
+public class GenerateReportCommand implements Command {
+
+    public void createReport() {
+        Scanner scanSatellite = new Scanner(System.in);
+        String criteria;
+        while (true) {
+            System.out.println("*** REPORT ***");
+            System.out.print("Enter the satellite or EXIT: ");
+            criteria = scanSatellite.next().trim().toUpperCase();
+            if (criteria.equals("EXIT"))
+                break;
+            try {
+                PDFReportFromFileBySatellite report = new PDFReportFromFileBySatellite(criteria);
+                report.generateReport();
+            } catch (Exception e) {
+                System.out.println("Error: Can't create the report! Try again!");
+            }
+        }
     }
 
     @Override
     public void doCommand() {
-        receiverDB.createReport();
+        createReport();
     }
 }
